Split Hermes discovery tests
